feat: Allow change routes in cli custom templates
